requirements.txt
alembic==1.14.1
annotated-types==0.7.0
fastapi==0.115.11
psycopg2==2.9.10
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.27.2
SQLAlchemy==2.0.38
typing_extensions==4.12.2
uvicorn==0.34.0
python-dotenv==1.0.1
apache-beam==2.63.0
asyncpg==0.30.0
asyncio==3.4.3
PyJWT==2.9.0
passlib==1.7.4
bcrypt==4.0.1
redis==5.2.1

# Google 
google-api-core==2.24.1
google-apitools==0.5.31
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.82.0
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.28.0
google-cloud-bigtable==2.29.0
google-cloud-core==2.4.2
google-cloud-datastore==2.20.2
google-cloud-dlp==3.28.0
google-cloud-language==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.16
google-cloud-resource-manager==1.14.1
google-cloud-secret-manager==2.23.1
google-cloud-spanner==3.52.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
grpc-google-iam-v1==0.14.1
keyrings.google-artifactregistry-auth==1.1.2
cloud-sql-python-connector==1.17.0
gcloud-aio-auth==5.3.2
gcloud-aio-storage==9.3.0
